Fix compatibility to symfony 4
